+++ src/main/java/techreborn/tiles/TilePlayerDectector.java
@@ -1,11 +1,11 @@
 package techreborn.tiles;
 
+import java.util.Iterator;
+
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraftforge.common.util.ForgeDirection;
 import techreborn.powerSystem.TilePowerAcceptor;
-
-import java.util.Iterator;
 
 
 public class TilePlayerDectector extends TilePowerAcceptor {
@@ -52,7 +52,7 @@ public class TilePlayerDectector extends TilePowerAcceptor {
                 Iterator tIterator = super.worldObj.playerEntities.iterator();
                 while (tIterator.hasNext()) {
                     EntityPlayer player = (EntityPlayer) tIterator.next();
-                    if (player.getDistanceSq((double) super.xCoord + 0.5D, (double) super.yCoord + 0.5D, (double) super.zCoord + 0.5D) <= 256.0D) {
+                    if (player.getDistanceSq(super.xCoord + 0.5D, super.yCoord + 0.5D, super.zCoord + 0.5D) <= 256.0D) {
                         if(worldObj.getBlockMetadata(xCoord, yCoord, zCoord) == 0){//ALL
                             redstone = true;
                         } else if (blockMetadata == 1){//Others
